// app/types/Parhaiyan.ts
export interface Parhaiyan {
    id?: number | string;           
    title_en?: string;              
    title_ur?: string;               
    description_en?: string;         
    description_ur?: string;         
    start_date?: string | Date;      
    end_date?: string | Date;        
    url_slug?: string;               
    is_active?: boolean;             
    created_at?: string | Date;      
    updated_at?: string | Date;      
  }
  